feat(accessctrl): add role checks to user module

+++ src/modules/user/routes.ts
@@ -3,6 +3,7 @@ import { UnAuthorizedException } from 'http-exception-transformer/exceptions'
 import { check } from '../../services/roles/definitions'
 import { resolveRole } from '../../services/roles/resolver'
 import { SuccessToResponseMapper } from '../../services/util/response.transformer'
+import { LoginTokenData } from '../auth/interface'
 import { UserController } from './controller'
 import { CreateUserInterface } from './interface'
 
@@ -13,7 +14,7 @@ const router = express.Router()
  */
 router.get('/', async (req: Request, res: Response) => {
   if (!check.can(resolveRole(req)).readAny('profile').granted) {
-    throw new UnAuthorizedException('Not enough rights to list of users')
+    throw new UnAuthorizedException('Not enough rights to list of all users')
   }
   const data = await UserController.getAllUsers()
   res.json(SuccessToResponseMapper(data))
@@ -22,13 +23,26 @@ router.get('/', async (req: Request, res: Response) => {
 /** to fetch details of any given user */
 router.get('/:email', async (req: Request, res: Response) => {
   const { email } = req.params
-  const data = await UserController.getUser(email)
-  res.json(SuccessToResponseMapper(data))
+  const cookieData: LoginTokenData = req.body.cookie
+
+  if (
+    check.can(resolveRole(req)).readAny('profile').granted ||
+    (check.can(resolveRole(req)).readOwn('profile').granted && email === cookieData.email)
+  ) {
+    const data = await UserController.getUser(email)
+    res.json(SuccessToResponseMapper(data))
+  }
+
+  throw new UnAuthorizedException('Not enough rights to list user details')
 })
 
 /** to create a new user */
-router.post('/', async ({ body }, res: Response) => {
-  const userDetails: CreateUserInterface = body.payload
+router.post('/', async (req: Request, res: Response) => {
+  if (!check.can(resolveRole(req)).createOwn('profile').granted) {
+    throw new UnAuthorizedException('Not allowed to create a new profile')
+  }
+
+  const userDetails: CreateUserInterface = req.body.payload
   const data = await UserController.createUser(userDetails)
   res.json(SuccessToResponseMapper(data))
 })
@@ -36,8 +50,17 @@ router.post('/', async ({ body }, res: Response) => {
 /** to delete an old user */
 router.delete('/:email', async (req: Request, res: Response) => {
   const { email } = req.params
-  const data = await UserController.deleteUser(email)
-  res.json(SuccessToResponseMapper(data))
+  const cookieData: LoginTokenData = req.body.cookie
+
+  if (
+    check.can(resolveRole(req)).deleteAny('profile').granted ||
+    (check.can(resolveRole(req)).deleteOwn('profile').granted && cookieData.email === email)
+  ) {
+    const data = await UserController.deleteUser(email)
+    res.json(SuccessToResponseMapper(data))
+  }
+
+  throw new UnAuthorizedException('Not enough rights to delete this user')
 })
 
 export default router
